postoffice: keep a copy of the receiver to avoid race condition
If all receivers are dropped the channel is closed. You can use
`tx.subscribe()` to create a new one and the channel reopens, but any
messages sent before this cause a send error.

When one task opens the mailbox for sending, and then sends a message
before any task opens it for reading the sender gets an error. Instead
the mailbox will always keep one receiver open so that sends never fail.